requirements.txt
streamlit
pandas
playwright
google-generativeai
openai
openpyxl